Implement spawn via trampoline mechanism.
Create a subprocess, by means of fork and exec.
A simple executable is embedded in Rust compiled shared library

When a call to spawn is made, the executable file is written as
memfd.

If a call is made from a shared library code. The path to shared library
and the name of entry point symbol name is extracted and passed to the
executable.

The executable is executed, and loads the shared library via dlopen(2)
Then looks up the symbol and runs it